fix: 使用类型报错为题
