+++ src/components/ContextMenu/index.tsx
@@ -1,67 +1,15 @@
-import React, { useState, useMemo, useEffect } from 'react';
-import classNames from 'classnames';
-import { useScene } from '../LarkMap/hooks/use-scene';
-import { Marker } from '../Marker';
-import { CLS_PREFIX } from './constant';
+import type React from 'react';
 import type { ContextMenuProps } from './types';
 import { ContextMenuItem } from './ContextMenuItem';
+import { ContextMenu as InternalContextMenu } from './ContextMenu';
 import './index.less';
 
-const ContextMenu: React.FC<ContextMenuProps> = (props) => {
-  const scene = useScene();
-  const [initMenu, setInitMenu] = useState({
-    visible: false,
-    position: undefined,
-  });
+export interface InternalContextMenuType extends React.FC<ContextMenuProps> {
+  Item: typeof ContextMenuItem;
+}
 
-  const initMenuRender = useMemo(() => {
-    return (
-      <div className={classNames(`${CLS_PREFIX}`, props?.className)} style={props?.style}>
-        {props?.children}
-      </div>
-    );
-  }, [props]);
+const ContextMenu = InternalContextMenu as unknown as InternalContextMenuType;
 
-  // 右键打开面板信息
-  const mapRightMenuOpen = (e) => {
-    const { lng, lat } = e.lnglat;
-    setInitMenu({
-      visible: true,
-      position: { lng, lat },
-    });
-  };
-
-  // 单击事件关闭菜单
-  const mapRightMenuClose = () => {
-    const timeOut = setTimeout(() => {
-      if (timeOut) {
-        clearTimeout(timeOut);
-        setInitMenu({
-          visible: false,
-          position: undefined,
-        });
-      }
-    }, 0);
-  };
-
-  useEffect(() => {
-    if (scene) {
-      scene.on('contextmenu', mapRightMenuOpen);
-      scene.on('click', mapRightMenuClose);
-    }
-    return () => {
-      scene.off('contextmenu', mapRightMenuOpen);
-      scene.off('click', mapRightMenuClose);
-    };
-  }, [scene]);
-
-  return initMenu.visible ? (
-    <Marker lngLat={initMenu.position} anchor="top-left">
-      {initMenuRender}
-    </Marker>
-  ) : null;
-};
-
-(ContextMenu as any).Item = ContextMenuItem;
+ContextMenu.Item = ContextMenuItem;
 
 export { ContextMenu };
